fix(Reward): resolved issues related to migrations

+++ rest/database/migrations/2020_07_02_052659_create_rewards_table.php
@@ -15,7 +15,7 @@ class CreateRewardsTable extends Migration
     {
         Schema::create('rewards', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('slot _id');
+            $table->unsignedBigInteger('slot _id');
             $table->string('rewardee_wallet');
             $table->decimal('reward_amount', 36, 18);
             $table->dateTime('rewarded_on');
